exports(5): update and correct information about subdirectory exports
Document that the default option is now no_subtree_check and add a
reference to the Subdirectory Exports section.

Add a warning to the Subdirectory Exports section that it is possible to
also access files on other filesystems based on a previous discussion.

Fix a typo in the Subdirectory Exports section. The correct option to
prevent access to files outside the subdirectory is subtree_check, not
no_subtree_check.

Signed-off-by: Philipp Tekeser-Glasz <philipp.tekeser-glasz@hvs-consulting.de>
Signed-off-by: Steve Dickson <steved@redhat.com>